requirements.txt
streamlit>=1.32.0
pdfplumber>=0.10.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langchain-core>=0.3.0
langchain-text-splitters>=0.3.0
openai>=1.50.0
faiss-cpu>=1.7.4
tiktoken>=0.7.0
bcrypt>=4.0.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
reportlab>=4.0.0
pillow>=10.0.0
